# api/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "A fitness API that provides workout routines and related nutrition\
    \ information."
  version: "1.0.0"
  title: "EpicLand API"
  contact:
    name: "MPM FITNIT"
host: "localhost:10010"
basePath: "/"
schemes:
- "http"
consumes:
- "application/json"
produces:
- "application/json"
- "text/plain"
paths:
...
  /users:
    # our controller name
    x-swagger-router-controller: "users"
    post:
      description: "Creates a new user in the database"
      operationId: "addNewUser"
      produces:
      - "application/json"
      parameters:
      - name: "user"
        in: "body"
        description: "new user to add to the database"
        required: true
        schema:
          type: "object"
          required:
          - "first_name"
          - "last_name"
          - "password"
          - "weight"
          - "user_intentions"
          properties:
            first_name:
              type: "string"
            last_name:
              type: "string"
            email:
              type: "string"
            password:
              type: "string"
            weight:
              type: "integer"
              format: "int32"
            user_intentions:
              type: "string"
        #   $ref: "#/definitions/newUser"
      responses:
        200:
          description: "New user response"
          schema:
            $ref: "#/definitions/user"
        400:
          description: "You didn't make a user"
          schema:
            $ref: "#/definitions/errorModel"
        default:
          description: "unexpected error"
          schema:
            $ref: "#/definitions/errorModel"
      x-swagger-router-controller: "users"
  /users/{id}:
    get:
      description: "Returns a user information based on a single ID. The user must\
        \ be authorized to access."
      operationId: "getUserById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of the user to fetch"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "user-specific response"
          schema:
            $ref: "#/definitions/user"
        default:
          description: "unexpected error"
          schema:
            $ref: "#/definitions/errorModel"
      x-swagger-router-controller: "users"
    delete:
      description: "deletes a single user from the database."
      operationId: "deleteUser"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of the user to delete"
        required: true
        type: "integer"
        format: "int32"
      responses:
        204:
          description: "user deleted"
        default:
          description: "unexpected error"
          schema:
            $ref: "#/definitions/errorModel"
      x-swagger-router-controller: "users"
    patch:
      description: "updates specified properties of the specific user."
      operationId: "patchUserProfile"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of specified user"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "jsonPatch"
        required: true
        schema:
          $ref: "#/definitions/updateUser"
      responses:
        200:
          description: "user profile response"
          schema:
            $ref: "#/definitions/user"
        default:
          description: "unexpected error"
          schema:
            $ref: "#/definitions/errorModel"
      x-swagger-router-controller: "users"
...
definitions:
...
  newUser:
    type: "object"
    required:
    - "email"
    - "first_name"
    - "last_name"
    - "password"
    - "weight"
    - "user_intentions"
    properties:
      # id:
      #   type: "integer"
      #   format: "int64"
      first_name:
        type: "string"
      last_name:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      weight:
        type: "integer"
        format: "int32"
      user_intentions:
        type: "string"
  user:
    type: "object"
    required:
    - "email"
    - "first_name"
    - "id"
    - "last_name"
    - "hashed_password"
    - "weight"
    - "user_intentions"
    properties:
      id:
        type: "integer"
        # format: "int64"
      first_name:
        type: "string"
      last_name:
        type: "string"
      email:
        type: "string"
      hashed_password:
        type: "string"
      weight:
        type: "integer"
        format: "int32"
      user_intentions:
        type: "string"
  updateUser:
    type: "object"
    required:
    - "password"
    properties:
      first_name:
        type: "string"
      last_name:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      weight:
        type: "integer"
        format: "int32"
      user_intentions:
        type: "string"
...
  errorModel:
    type: "object"
    required:
    - "code"
    - "message"
    properties:
      code:
        type: "integer"
        format: "int32"
      message:
        type: "string"